+++ sys12_gcc/funcs.h
@@ -1,3 +1,3 @@
-int arrayAvg(int *aryP,int aryLength);
+double arrayAvg(int *aryP,int aryLength);
 void copyArray(int *a0, int *a1, int size);
 int strLength(char *str);
